Refactored Package structure.
