+++ build.gradle
@@ -17,9 +17,6 @@ jacocoTestReport {
     }
 }
 
-dependencies {
-}
-
 allprojects {
     apply plugin: 'java'
     apply plugin: 'com.github.sherter.google-java-format'
@@ -68,7 +65,7 @@ allprojects {
     sourceSets {
         main {
             resources {
-                srcDirs = ["src/main/resource", "../shared-resources"]
+                srcDirs = ["src/main/resources", "../shared-resources"]
             }
         }
     }
