+++ owltorrent-common/README.md
@@ -1,0 +1,2 @@
+The commons package should be used to write the entity and 
+interfaces that would be used across packages, and also some
