+++ owltorrent-core/README.md
@@ -1,3 +1,1 @@
-The core package should be used to write the most essential 
-functionalities, such as the torrent protocol, and how to read
-a torrent file.
+The core package should be used to bring the different components together.
