--- owltorrent-locator/README.md
@@ -1,2 +1,0 @@
-The locator package should be used to write the implementation of how
-to find the first seeders, either using a tracker or a DHT.
