--- owltorrent-locator/build.gradle
@@ -1,6 +1,0 @@
-plugins {
-}
-
-dependencies {
-    implementation project(":owltorrent-core")
-}
