--- owltorrent-locator/src/main/java/edu/rice/owltorrent/locator/Temp.java
@@ -1,3 +1,0 @@
-package edu.rice.owltorrent.locator;
-
-public class Temp {}
