+++ owltorrent-network/README.md
@@ -1,0 +1,3 @@
+The network package should be used to write the implementation of network related
+classes, such as how to talk to peers, how to find the first seeders (using a 
+tracker or a DHT), etc.
