+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'owltorrent'
 include 'owltorrent-core'
 include 'owltorrent-client'
-include 'owltorrent-locator'
+include 'owltorrent-network'
 include 'owltorrent-storage'
+include 'owltorrent-common'
